+++ src/Products/PloneMeeting/browser/overrides.py
@@ -437,7 +437,7 @@ class PMRenderTermView(RenderTermPortletView):
 
     def __call__(self, term, category, widget):
         rendered_term = super(PMRenderTermView, self).__call__(term, category, widget)
-        # display the searchallmeetings as a selection list
+        # display the searchallmeetings/searchlastdecisions as a selection list
         if self.context.getId() in ['searchallmeetings', 'searchlastdecisions']:
             rendered_term = "<div id='async_search_term_{0}' class='loading' data-collection_uid='{0}'>" \
                 "<img src='{1}/spinner_small.gif' /></div>".format(